refact: components moved to src dir
