+++ src/components/RoomGrid.tsx
