+++ src/components/editor/DocContent.tsx
