+++ src/components/editor/EditorContent.tsx
