+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export * from './layouts/MainLayout'
+export * from './PostCard'
+export * from './PostContainer'
+export * from './Providers'
